requirements.txt
# VisionMoto v2.0 - Challenge 2025
# Dependências principais

# Visão Computacional
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Processamento
filterpy>=1.4.0
scipy>=1.10.0

# Backend API
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
requests>=2.31.0
gunicorn>=21.2.0

# IoT & Comunicação
paho-mqtt>=1.6.0

# Segurança & Auth
pyjwt>=2.8.0
werkzeug>=2.3.0

# Validação de Dados
pydantic>=2.0.0
email-validator>=2.0.0

# Banco de Dados
psycopg2-binary>=2.9.0
redis>=4.6.0

# Desenvolvimento & Testes
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0